reorg of build_requires tests
