+++ conans/test/functional/build_requires/build_requires_test.py
@@ -51,6 +51,49 @@ nonexistingpattern*: SomeTool/1.2@user/channel
 
 class BuildRequiresTest(unittest.TestCase):
 
+    def create_with_tests_and_build_requires_test(self):
+        client = TestClient()
+        # Generate and export the build_require recipe
+        conanfile = """from conans import ConanFile
+class MyBuildRequire(ConanFile):
+    def package_info(self):
+        self.env_info.MYVAR="1"
+"""
+        client.save({"conanfile.py": conanfile})
+        client.run("create . Build1/0.1@conan/stable")
+        client.save({"conanfile.py": conanfile.replace('MYVAR="1"', 'MYVAR2="2"')})
+        client.run("create . Build2/0.1@conan/stable")
+
+        # Create a recipe that will use a profile requiring the build_require
+        client.save({"conanfile.py": """from conans import ConanFile
+import os
+
+class MyLib(ConanFile):
+    build_requires = "Build2/0.1@conan/stable"
+    def build(self):
+        assert(os.environ['MYVAR']=='1')
+        assert(os.environ['MYVAR2']=='2')
+
+""", "myprofile": '''
+[build_requires]
+Build1/0.1@conan/stable
+''',
+                    "test_package/conanfile.py": """from conans import ConanFile
+import os
+
+class MyTest(ConanFile):
+    def build(self):
+        assert(os.environ['MYVAR']=='1')
+    def test(self):
+        self.output.info("TESTING!!!")
+"""}, clean_first=True)
+
+        # Test that the build require is applyed to testing
+        client.run("create . Lib/0.1@conan/stable --profile=./myprofile")
+        self.assertEqual(1, str(client.out).count("Lib/0.1@conan/stable: Applying build-requirement:"
+                                                  " Build1/0.1@conan/stable"))
+        self.assertIn("TESTING!!", client.user_io.out)
+
     def test_dependents_txt(self):
         client = TestClient()
         boost = """from conans import ConanFile
