+++ conans/test/unittests/model/transitive_reqs_test.py
@@ -430,7 +430,7 @@ class ChatConan(ConanFile):
         self.retriever.conan(bye_ref, bye_content2)
 
         with self.assertRaisesRegexp(ConanException, "Conflict in Bye/0.2@user/testing"):
-            deps_graph = self.root(chat_content)
+            _ = self.root(chat_content)
 
     def test_diamond_conflict_solved(self):
         chat_content = """
@@ -1320,7 +1320,7 @@ class HelloConan(ConanFile):
         self.assertEqual(hello.conanfile.info.full_requires.dumps(),
                          "Say/0.1@user/testing:%s" % NO_SETTINGS_PACKAGE_ID)
 
-    def test_remove_build_requires(self):
+    def test_remove_requires(self):
         hello_content = """
 from conans import ConanFile
 
@@ -1343,7 +1343,7 @@ class HelloConan(ConanFile):
         self.assertEqual(hello.conanfile.info.full_requires.dumps(),
                          "Say/0.1@user/testing:%s" % NO_SETTINGS_PACKAGE_ID)
 
-    def test_remove_two_build_requires(self):
+    def test_remove_two_requires(self):
         chat_content = """
 from conans import ConanFile
 
@@ -1609,7 +1609,8 @@ class LibDConan(ConanFile):
         libc_ref = ConanFileReference.loads("LibC/0.1@user/testing")
         self.retriever.conan(libc_ref, libc_content)
 
-        with self.assertRaisesRegexp(ConanException, "LibD/0.1@user/testing tried to change LibB/0.1@user/testing "
+        with self.assertRaisesRegexp(ConanException,
+                                     "LibD/0.1@user/testing tried to change LibB/0.1@user/testing "
                                      "option LibA:shared to True"):
             self.root(self.consumer_content)
